Workaround issues with Duro cell modem not loading
Compile usb support as a modules so that we can load them dynamically
and augment the cell_modem_daemon service such that we can detect if the
USB cell modem failed to enumerate and retry.

+++ board/piksiv3/rootfs-overlay/etc/init.d/logging.sh
@@ -1,18 +1,19 @@
-[[ -n "$log_tag" ]] || { echo "ERROR: 'log_tag' variable is not defined" >&2; exit 1; }
-
 log_fac=daemon
 
 log_base()
 {
+  [[ -n "$log_tag" ]] || { echo "ERROR: 'log_tag' variable is not defined" >&2; exit 1; }
+
   local send_sbp=
   while [[ $# -gt 0 ]]; do
     case "$1" in
       --sbp) send_sbp=y; shift;;
+      -s) send_stderr="-s"; shift;;
       *) break;;
     esac
   done
   [[ -z "$send_sbp" ]] || { echo $* | sbp_log --"${_log_level}"; }
-  logger -t $log_tag -p ${log_fac}.${_log_level} $*;
+  echo $* | logger -t $log_tag -p ${log_fac}.${_log_level} ${send_stderr:-}
 }
 
 logw() { _log_level=warn;  log_base $*; }
@@ -20,8 +21,10 @@ logi() { _log_level=info;  log_base $*; }
 logd() { _log_level=debug; log_base $*; }
 loge() { _log_level=err;   log_base $*; }
 
-setup_loggers()
+_setup_loggers()
 {
+  [[ -n "$log_tag" ]] || { echo "ERROR: 'log_tag' variable is not defined" >&2; exit 1; }
+
   logger_stdout=/var/run/$log_tag.stdout
   rm -f $logger_stdout
   mkfifo $logger_stdout
@@ -46,5 +49,8 @@ cleanup_loggers()
   rm ${logger_stdout} ${logger_stderr}
 }
 
-setup_loggers
-trap 'cleanup_loggers' EXIT TERM
+setup_loggers()
+{
+  _setup_loggers
+  trap 'cleanup_loggers' EXIT TERM
+}
